+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## feat
 
 - add robot-message-spring-boot-starter; 钉钉机器人, 企业微信机器人, 邮箱, 短信
+- excel-spring-boot-starter
+- netty-spring-boot-starter
+
 # 2.2.3-Bata
 
 - add push-spring-boot-starter